requirements.txt
appnope==0.1.0
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.4
django-extensions==2.2.9
djangorestframework==3.11.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
